Some extra overview formatting

+++ yogh-explorer-client/src/main/java/nl/yogh/wui/explorer/service/domain/TransactionSummary.java
@@ -6,7 +6,7 @@ import jsinterop.annotations.JsType;
 @JsType
 public class TransactionSummary {
   @JsProperty public String txid;
-  @JsProperty public long fee;
-  @JsProperty public long vsize;
-  @JsProperty public long value;
+  @JsProperty public int fee;
+  @JsProperty public int vsize;
+  @JsProperty public int value;
 }
